requirements.txt
cairocffi==0.8.0
CairoSVG==1.0.22
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
Django==1.11.10
django-bootstrap3==9.1.0
django-crispy-forms==1.7.1
django-dbbackup==3.2.0
django-easy-pdf==0.1.1
django-filter==1.1.0
django-tables2==1.19.0
djangorestframework==3.7.7
et-xmlfile==1.0.1
gnupg==2.3.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
jdcal==1.3
odfpy==1.3.6
openpyxl==2.5.0
pdfrw==0.4
Pillow==5.0.0
psutil==5.4.3
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
six==1.11.0
tablib==0.12.1
tinycss2==0.6.1
unicodecsv==0.14.1
urllib3==1.22
WeasyPrint==0.42.2
webencodings==0.5.1
xhtml2pdf==0.2.1
xlrd==1.1.0
xlwt==1.3.0
